+++ src/main/java/com/concurrency/thread/counter/Counter.java
@@ -1,0 +1,9 @@
+package com.concurrency.thread.counter;
+
+public interface Counter {
+    void add(int value);
+
+    int show();
+
+    void clear();
+}
