+++ src/main/java/com/concurrency/thread/counter/batch/BatchCounter.java
@@ -1,0 +1,7 @@
+package com.concurrency.thread.counter.batch;
+
+import com.concurrency.thread.counter.Counter;
+
+public interface BatchCounter extends Counter {
+    void flush();
+}
