+++ src/main/java/com/concurrency/thread/counter/queue/Consumer.java
@@ -1,0 +1,12 @@
+package com.concurrency.thread.counter.queue;
+
+import org.springframework.context.annotation.Profile;
+
+import java.util.concurrent.TimeUnit;
+
+@Profile("dev")
+public interface Consumer {
+    void consumeEvent(long timeout, TimeUnit unit) throws InterruptedException;
+
+    Long show();
+}
