+++ src/main/java/com/concurrency/thread/counter/queue/Producer.java
@@ -1,0 +1,8 @@
+package com.concurrency.thread.counter.queue;
+
+import org.springframework.context.annotation.Profile;
+
+@Profile("dev")
+public interface Producer {
+    void add(long value) throws InterruptedException;
+}
