+++ src/main/java/com/concurrency/thread/executor/Performance.java
@@ -1,0 +1,8 @@
+package com.concurrency.thread.executor;
+
+public record Performance(String name, long time, int threads, long memory) {
+    @Override
+    public String toString() {
+        return String.format("| %-20s | %10d ms | %5d threads | %10d KB |", name, time, threads, memory / 1024);
+    }
+}
