--- src/main/java/com/thread/concurrency/counter/Counter.java
@@ -1,9 +1,0 @@
-package com.thread.concurrency.counter;
-
-public interface Counter {
-    void add(int value);
-
-    int show();
-
-    void clear();
-}
