--- src/main/java/com/thread/concurrency/counter/batch/BatchCounter.java
@@ -1,7 +1,0 @@
-package com.thread.concurrency.counter.batch;
-
-import com.thread.concurrency.counter.Counter;
-
-public interface BatchCounter extends Counter {
-    void flush();
-}
