--- src/main/java/com/thread/concurrency/counter/queue/Consumer.java
@@ -1,12 +1,0 @@
-package com.thread.concurrency.counter.queue;
-
-import org.springframework.context.annotation.Profile;
-
-import java.util.concurrent.TimeUnit;
-
-@Profile("dev")
-public interface Consumer {
-    void consumeEvent(long timeout, TimeUnit unit) throws InterruptedException;
-
-    Long show();
-}
