--- src/main/java/com/thread/concurrency/executor/Performance.java
@@ -1,8 +1,0 @@
-package com.thread.concurrency.executor;
-
-public record Performance(String name, long time, int threads, long memory) {
-    @Override
-    public String toString() {
-        return String.format("| %-20s | %10d ms | %5d threads | %10d KB |", name, time, threads, memory / 1024);
-    }
-}
