--- src/test/java/com/thread/concurrency/counter/SpringWebJpaConcurrencyApplicationTests.java
@@ -1,11 +1,0 @@
-package com.thread.concurrency.counter;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringWebJpaConcurrencyApplicationTests {
-    @Test
-    void contextLoads() {
-    }
-}
